Updated how NodeElement Operations Behave.
 Changes to be committed:
	modified:   examples/example.rb
	modified:   lib/bufs_base_node.rb
	modified:   lib/bufs_node_factory.rb
	modified:   lib/midas/node_element_operations.rb
	deleted:    spec/needs_updated_node_element_operations_spec.rb
	new file:   spec/node_element_operations_spec.rb
Node Element Operations is now a separate class and each User class will have a single instace of it.
Also, specs were created for it and behavior tested and verified.
Also, IMPORTANT: bufs_base_node was updated to inlcude a method_missing that checks to see if a method coming in looks like it should be applied to the user data. If so, then the data is returned as is. If not, then a method missing error is generated.