query all is confusing
Fixing the confusion (this commit is in the middle of the fixing so
its very disorganized).
There are two types of "all" querying.
.all => returns all tinkit nodes from persistence layer
.query_all => returns native persistence layer records

.query_all needs to have a better name (.native_query_all?)

In additon the .all query has some sugar that will let you add or
subtract fields on the fly, and has a common algorithm for processing
the native data. The problem is that I'm not sure a common algorithm
will work in all cases, and that it would be better to have a method
in the glue to pre-process the data into a common format (hashes like
the nodes are initialized with). So the algo and the modification of
fields both need to be taken into consideration.
Also, the spec tests are not very robust.